requirements.txt
python-telegram-bot>=20.7
python-dotenv>=1.0.0
requests>=2.31.0
google-generativeai>=0.3.0
Pillow>=10.0.0
python-magic>=0.4.27 